--- __mocks__/openmrs-esm-api.mock.tsx
@@ -1,15 +1,0 @@
-import { never, of } from "rxjs";
-
-export function openmrsFetch() {
-  return new Promise(() => {});
-}
-
-export function openmrsObservableFetch() {
-  return of({ data: { entry: [] } });
-}
-
-export function getCurrentPatient() {
-  return jest.fn().mockReturnValue(never());
-}
-
-export const fhirBaseUrl = "/ws/fhir2/R4";
