--- __mocks__/openmrs-esm-config.mock.tsx
@@ -1,8 +1,0 @@
-export function defineConfigSchema() {}
-
-export const validators = {
-  isBoolean: jest.fn(),
-  isString: jest.fn(),
-  isUuid: jest.fn(),
-  isObject: jest.fn()
-};
