--- __mocks__/openmrs-esm-extensions.mock.tsx
@@ -1,3 +1,0 @@
-import React from "react";
-
-export const switchTo = jest.fn();
