--- __mocks__/openmrs-esm-styleguide.mock.tsx
@@ -1,1 +1,0 @@
-export const showToast = jest.fn();
