+++ jest.config.json
@@ -8,12 +8,7 @@
     "\\.(jpg|jpeg|png)$": "<rootDir>/__mocks__/file.mock.ts",
     "^@carbon/icons-react/es/(.*)$": "@carbon/icons-react/lib/$1",
     "^carbon-components-react/es/(.*)$": "carbon-components-react/lib/$1",
-    "@openmrs/esm-api": "<rootDir>/__mocks__/openmrs-esm-api.mock.tsx",
-    "@openmrs/esm-error-handling": "<rootDir>/__mocks__/openmrs-esm-error-handling.mock.tsx",
-    "@openmrs/esm-config": "<rootDir>/__mocks__/openmrs-esm-config.mock.tsx",
-    "@openmrs/esm-react-utils": "<rootDir>/__mocks__/openmrs-esm-react-utils.mock.tsx",
-    "@openmrs/esm-styleguide": "<rootDir>/__mocks__/openmrs-esm-styleguide.mock.tsx",
-    "@openmrs/esm-extensions": "<rootDir>/__mocks__/openmrs-esm-extensions.mock.tsx",
+    "@openmrs/esm-framework": "<rootDir>/__mocks__/openmrs-esm-framework.mock.tsx",
     "@carbon/charts/interfaces/enums":"<rootDir>/__mocks__/carbon-interface.ts",
     "^lodash-es/(.*)$": "lodash/$1"
   },
