+++ src/widgets/visit/new-visit.component.test.tsx
@@ -14,11 +14,8 @@ import NewVisit from "./new-visit.component";
 const mockGetCurrentPatientUuid = getCurrentPatientUuid as jest.Mock;
 const mockOpenmrsObservableFetch = openmrsObservableFetch as jest.Mock;
 
-jest.mock("@openmrs/esm-framework", () => ({
-  getCurrentPatientUuid: jest.fn(),
-  openmrsObservableFetch: jest.fn(),
-  fhirBaseUrl: "/ws/fhir2"
-}));
+mockGetCurrentPatientUuid.mockImplementation(jest.fn());
+mockOpenmrsObservableFetch.mockImplementation(jest.fn());
 
 describe("<NewVisit />", () => {
   let patientUuid = "some-patient-uuid";
