+++ src/widgets/visit/start-visit-prompt.component.test.tsx
@@ -1,11 +1,6 @@
 import { startVisitPrompt } from "./start-visit-prompt.component";
-import { newModalItem } from "./visit-dialog.resource";
 
-jest.mock("./visit-dialog.resource", () => ({
-  newModalItem: jest.fn()
-}));
-
-const mockNewModalItem = newModalItem as jest.Mock;
+const mockNewModalItem = jest.fn();
 
 describe("StartVisitPrompt", () => {
   it("should open StartVisit Prompt", () => {
