+++ src/widgets/vitals/vitals-biometrics-form/vitals-biometrics-form.component.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { screen, render, act } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import VitalsAndBiometricForms from "./vitals-biometrics-form.component";
+import { screen, render, act } from "@testing-library/react";
 import { savePatientVitals } from "../vitals-biometrics.resource";
 import { mockSessionDataResponse } from "../../../../__mocks__/session.mock";
 import { mockVitalsSignsConcept } from "../../../../__mocks__/vitals.mock";
@@ -13,10 +13,8 @@ const mockOpenmrsFetch = openmrsFetch as jest.Mock;
 const mockSavePatientVitals = savePatientVitals as jest.Mock;
 const mockCloseWorkspace = jest.fn();
 
-jest.mock("@openmrs/esm-framework", () => ({
-  openmrsObservableFetch: jest.fn(),
-  openmrsFetch: jest.fn()
-}));
+mockOpenmrsObservableFetch.mockImplementation(jest.fn());
+mockOpenmrsFetch.mockImplementation(jest.fn());
 
 jest.mock("../vitals-biometrics.resource", () => ({
   savePatientVitals: jest.fn()
